Dev inv

+++ ECU/Inc/devices/torquecontrol.h
@@ -71,7 +71,7 @@ int getTorqueReqPercR( int16_t pedalreq );
 
 
 
-float PedalTorqueRequest( int16_t *used_pedal_percent );
+int PedalTorqueRequest( int16_t *used_pedal_percent );
 
 void SetupTorque( uint8_t pedal );
 bool SetupInterpolationTables( eepromdata* eepromdatahandle );
